Initial commit.
Bunch of forms were generated using the form maker and the database. Forms are still rough and need to be adjusted (some data for the form should not be fields, but simply provided by the form (like the clients's ids, the available contracts, etc.

session classes/session_class.php is in progress. Pretty much all that's needed to load a session. and each key method is covered by a test.